Fixed some things in this English Article
probably the most ugliest around here. Fixed some to Markdown

still WIP

unfortunately because MD is limited by design, I cannot unuse some HTML around here such as the font size

which is according to [this stack exchange topic](http://meta.stackexchange.com/questions/170480/is-it-possible-to-change-the-text-size-in-the-code-markdown). the text sizing is entirely up to the engine itself.