Added more subheaders. Linebreaking. Some text changes. Moving paragraph
Specifically, the player box was moved down to the Team Mode as part of General section. Lazy effort to check and change for passive writing. Changing some old lifebar terms to health bar. Random text fixes and linebreaking for long paragraph that using a subheader look odd and no clue what to name it.