v1 folder for all main Game_Modes folders.
Almost carbon copy of the current page version with minor changes at the top. The notice is stated that the v1 page is based on the old ScoreV1 system used. The top link was also changed to link to its own current version rather than the Game_Modes index page. Expected for writers to directly update the page if ScoreV2 actually becomes the new standard (and having a link at top to link to v1).