Beatmap Nominator Rules: Recent changes
As result of discussion, among both the QAT as well as the BNs, we have changed the 24 hour rule to ignore renominations for minor changes, due to the recently introduced self-popping mechanic in modding discussions. This way there is still an opportunity to veto, and there is also more of an incentive to correct minor issues before qualification.

This follows the spirit of the "**Do not qualify a beatmap if any significant changes were made since it was bubbled**" rule as it was before modding discussions were introduced.