Ranking Criteria: Some minor adjustments
Removed "Difficulties lower than the hardest difficulty of a beatmapset can use their play time as a metric instead of drain time, but their drain time must be equal to at least 80% of their play time. (Not applicable to difficulties below 30 seconds of drain time.)" under the new spread rules

Changed "**If the drain time of a beatmap is...**" to "**If the drain time of the highest difficulty is...**" to basically implement the same thing as above but in an easier to read and less convoluted way

Removed the pural of "any guest difficulty beatmap creators", makes it less likely to be confused with all in total.