New Changes to Star Rating and PP
"Hello!

It is with anticipation that I, Xexxar, am here to announce changes to both the PP system as well as the star rating system. In fact, the star rating system, formally known as DiffCalc, has not received an update in several years, making this one of the most exciting updates to the game in a long time. The game has changed considerably since the last update to DiffCalc with the introduction of many new styles of maps and gameplay. While this change is not designed to fix every issue with the system, it's fundamental purpose is to better assess skills that may have been missing from previous iterations of the algorithm. Without further delay, let's get into the specifics.

## Wide Angle Jumps

As many players are well aware, the angle of a jump impacts how difficulty a pattern is to combo. Evidence of this can be found all throughout modern mapping. Almost all maps that are popular for their PP values contain jumps that are typically non-linear. As such, through the usage of a new angle calculation, a new scaling reward for jumps with angles exceeding 60 degrees has been added. Decided not to farm a map because it had square jumps? Found yourself baffled by the difficulty of an older 2013 map with linear flow? These previously underweighted maps now have more accurate star rating associated with them (this includes alt maps). And yes, they will be worth more PP as a result.

## Streams and High BPM

Besides the changes to aim, there has been an adjustment to how streams are handled by the algorithm. Many players who are capable of incredible speeds find their ranks are unfairly low given their unique skillset. In order to promote these unique skillsets, streams exceeding 200 BPM now have an increased reward that more accurately represents their difficulty. Able to FC 300 BPM stacks with your laptop keyboard? This is the change you've been waiting for. 

In addition, a small nerf to very spaced streams has been introduced. These are those streams found in maps such as Honesty and Sidetracked Days. While the nerf is sizable, these plays are still quite impressive and are still worth a lot, just not quite as much as they were worth before.

## Flashlight Reworks

One of the exciting introductions that will be coming with this patch is the change to flashlight. Primarily a dead mod outside of farming 30 second maps, flashlight has been irrelevant since PPV2's inception back in 2014. In order to remedy this, flashlight will now scale at a much higher rate with the length of a map. To compensate for this increased buff for longer maps, the shorter maps that were previously played with flashlight will receive a small nerf.

## Hidden Rework

Yes, HD is changing again. With the introduction of the high BPM buff, we wanted to make changes to HD so that maps with fast stacked streams weren't overly buffed when HD is applied. Instead of the old system, HD now rewards accuracy more, and both aim and speed PP will scale off of the approach rate of the map. Find yourself good at farming EZHD plays? You'll find a notable increase in performance points for these plays after the update. 

## Misc PP Changes

These are the more minor technical changes that we introduced to keep things balanced. In order to combat potentially unimpressive keyboard mashing that may be common for high speed plays as a result of the new buff, a new reward curve for accuracy has been introduced for the speed skill. Note that this will only punish the speed PP from plays with accuracy under 95%. Aim is unaffected, so your 90% Harumachi Clover FC should still be worth around what it was before. 

Previously the algorithm rewarded extra aim PP for plays that were AR10.33+. While we left this in, we reduced the reward and instead introduced the same concept for speed PP as well. This means that impressive HRDT speed plays will now receive a more accurate amount of pp given their difficulty.




And that's it! These long needed changes to your scores will roll out over the next few days as scores are being reprocessed. I hope that these changes shape up the leaderboard in a way that is both fun and exciting for all players. Stay tuned for more updates in the future and happy farming!"





Hopefully this works. Please let me know what I can improve below.