make stylistic and grammar edits
a few changes in the way rules are phrased to clarify them and make them sound more natural
this is by no means final, and there may be more grammatical issues that may need to be fixed

things i noticed but didn't change: pick one term for map pool, map set, etc. and stick with it
is tournament management capitalized? both are used throughout
there's a lot of superfluous words throughout the whole article; i figured removing all of them would be too radical but try to write tight if possible