Fixing even more typos and applying suggestions
A few typo fixes aside from the suggestions made, clarified comprehension in a few lines. This should be good to go now.

Unsure if all of the suggested changes should be individually marked as resolved to avoid spam, but if it is like so, all of those have been resolved in this commit.