[PT-BR ] Update Shortcut_key_reference
Created an entry to the pt-br translation for all of the shortcut combinations.

Note for native reviewers: Please check if the translation of a few terms vs keeping a few terms in English (especially mapping terms) is acceptable. If not, feel free to comment and point them out so that I may update them.