Start of my dutch translation of the wiki
The start of my translation of the wiki to dutch, also includes fixes some minor mistakes in the english pages.
Note: part of the interface is outdated so I updated it.
A different part that is also outdated (final part of interface) will probably need some updating aswell, since the results screen changed, but I left this alone for now.

Second note: I translated the wiki to dutch except for a couple words/sentences. I think referring to groups ,mods and a handful of other things in their english variant sounds a lot better and makes more sense. I might want to change things of the game to dutch as the local leaderboards and a variety of other sections got translated.