Interface update
Updated the interface page.
The old page had outdated versions of the bubble icon,  which got outdated due to the removal of forum modding/addition of the modding
section on beatmaps.
Note: this change was made in my previous request so idk how this would work.
Changed part of the delete all visible beatmaps as I think this is no longer a joke. Changed the part so that it keeps some history of this button.
Biggest change is updating the part of the online leaderboard. It is now up to date and shows the correct screen. May need a different screenshot though.