Update Interface
updates the interface article.
Fixes outdated parts and the same spelling mistakes as in my previous upload:
Fixes:
Bubble icon: Change due to modding page instead of forum modding.
Delete beatmaps: Not 100% you can do this but I don't want to test delete all visible maps with 105k maps. Changed the part so that the 'joke' is now seen as something in the past.
Online Ranking:
Got outdated after updates to the screen you get. Updated pretty much the entire part. Also added a new screenshot for this section, and deleted the old one.
May require a different screenshot, as mine isn't the best.