Cleanup Help Centre
This is a WIP in-depth cleanup of the Help Centre. Whether or not these sections get moved, this will make sure the spelling, grammar, punctuation, and styling stay squeaky-clean and as consistent as possible. I also added some extra info. here and there.

I have not messed with any contractions because I'm bot sure if this article should use them or not. Though I have replaced all contractions that use the word "not" with their two-word counterparts to add emphasis.

**Note for toaster:** You last edited **line 486**. So just read from "Allowing a program through Windows Firewall" to whatever on the next commit.

Things that toaster needs to double-check:
- [ ] Ctrl+F "sign in" and "in game" to make sure that the hyphens are placed where they should be
- [ ] Crtl+F all semicolons, hyphens, and ellispes to make sure that their uses are grammatically correct
- [ ] Ctrl+F "internet" to make sure that the "I" is capitalised
- [ ] Ctrl+F all number's symbols (1, 2, 3, 4, 5, 6, 7, 8, 9, 0) to make sure they are spelled-out when needed
- [ ] Ctrl+F all number's spelled-out variations under 11 (zero, one, two, three, four, five, six, seven, eight, nine, ten) to make sure they use the symbol if needed
-