[DE] Update Beginner's Tutorial
Never again, lol. So much time spent for the two people actually reading this in German ^^"

This PR is a general overhaul of the original article and should also no longer be outdated now. It's my first contribution to the wiki, so I'm not sure if I'm supposed to delete the flags in the beginning or if someone else does that. I guess "no_native_review" holds true until another German speaker has reviewed it?

To the German speakers: I translated the term "(graphics) tablet" as "Grafiktablett" once (the correct translation), but just used "Tablet" everywhere else as this is the term commonly used. I think that's a nice middle way, so please give feedback if you disagree.

Most probably there are still a few mistakes in it, I'll have another look at it tomorrow.