Rename and various changes
As suggested by clayton i think it's good to rename this issue to "fall", as November also almost ended. I thought about skipping November anyways and go on with December. With the re-name we avoid this "missing month" issue, as November counts to fall! 

Furthermore i changed all "This month" to "this issue" or "release".

Last but not least i almost forgot to add nepuris leave. Fixed that too!