Correcting an incorrect user
In the list August it states "-Added Tasha to osu!taiko Beatmap Appreciation Team" but the user that is being linked is incorrect. The correct user is karterfreak. https://osu.ppy.sh/users/1031958 This was explained clarified by OnosakiHito who was a founding member of the osu!Taiko BAT alongside karterfreak. https://imgur.com/myZSrUM