Update to reflect changes in timezones due to summer time ending
On 31st October (the last Sunday in the month) all countries in Europe turn the clocks back an hour. This means that some of the stated timezones and local times are incorrect. Daylight saving time also ends in Mexico for most regions. The one other country in the Sunday list that uses DST is Chile, but it ended there in early September.

Also, next Sunday, DST will end in Canada.