Replace all usage of "osu!lazer"
This is no longer correct terminology. "lazer" was always intended to be
a codename and has already been dropped from all public facing usage in
releases some time ago.

I am intentionally wanting to rewrite history here to avoid the name
sticking, since it'll either be no longer relevant or replaced with a
new release name in the future.

The `(lazer release)` suffix will eventually be dropped when things go
live, although that will need to be done alongside careful distinction
of articles that are specific to the old client behaviour.