Add guideline regarding long note length to o!m RC
This guideline proposal is the product of the initial RC Proposal in the osu! forum during the past November: https://osu.ppy.sh/community/forums/topics/1463447. 

The object of this proposal is to act as a baseline to indicate the minimum length warranted for long notes in osu!mania, following overwhelming community and BN/NAT support. 

We have determined that the presence of a long note release timing, together with possible limitations due to the hardware/software used, makes very short LNs (shorter than ~25-30ms) unreasonable to be hit and released with good accuracy even if they appear isolated, which makes them unsuitable for the ranked section.