osu!monthly #1
We are just waiting for 0x84f's osu!lazer article and our content will be complete for the month.
We wanted to start the review now, as we want to release it the 1st of August, and I have less time this week due to a travel.

Thanks to everybody who'll review !

Co-Authored-By: 0x84f <24534022+0x84f@users.noreply.github.com>
Co-Authored-By: Hubz <hubz@hubza.co.uk>
Co-Authored-By: wowcake <76740201+wowcake@users.noreply.github.com>
Co-Authored-By: Daniil <36848514+Kimitakari@users.noreply.github.com>
Co-Authored-By: Blooshing <90729492+Blooshing@users.noreply.github.com>