Clarify custom hitnormal obligatoriness in o!m hitsound rule exception
The goal of this PR is to add a simple clarification to the osu!mania exception of the "mandatory hitsounds" rule in General RC. 

While the current wording does not obligate mappers to add one or more custom hitnormal samples overriding the default hitnormal, it has for years been considered an unwritten rule or standard *de facto* to override the default hitnormal in all osu!mania ranked sets, in order to improve the audible feedback of the map for those users who prefer to enable hitsounds. 

Generally this is done to use less "punchy" hitnormals instead of the default ones, which usually become obnoxious to the player due to the typical high object density and concurrency in osu!mania beatmaps. Hitnormals used by mappers as overrides of course still provide acceptable audible feedback (this is mandated by RC, anyway) but they are less overwhelming and provide a better experience for those players who choose to play with hitsounds.

Thus, current RC reads: 


> **osu!mania beatmaps do not require hitsound additions.**[^normal-vs-addition] This is to allow for easier approachability to osu!mania mappers of different upbringings. It is still highly recommended to use hitsound additions to improve the feel of your beatmaps.

and after discussing among the osu!mania BN and NAT, we propose the following clarification, also simplifying the rule a bit:

> **osu!mania beatmaps do not require hitsound additions, but a custom hitnormal sample must override the default hitnormal.**[^normal-vs-addition] It is still recommended to use hitsound additions to improve the audible feedback of your beatmaps.

Of course, any proposal for an alternative or more clear wording is welcomed

Cheers