Change license summary link.
The summary previously provided by tldrlegal.com is only available in English and may be problematic when referencing it in translations. Creative Commons, the organization behind the licenses of the same name, already provides their own summary that provides basically the same information and is already translated into other languages.