Update ko.md
원본 리더십 항목 56번째 줄에서 `Inactive NAT members will be confronted by the [team leaders](#nat-leadership). If a timely resolution to their inactivity is not feasible, they will be removed from the NAT. Members under the evaluation category working on other mapping-related projects may be moved to the structural category to better reflect their contribution.`를 

`Each month, the NAT members are required to submit a summary of their activity in the [BN/NAT website](https://bn.mappersguild.com/). This summary, alongside other metrics like nomination and evaluation activity, is used to determine whether a member is still active and whether they should remain in the group. The [team leaders](#nat-leadership) will then discuss the activity of each member and decide whether they should remain in the NAT.

The team leaders will confront inactive NAT members, or members who fail to provide a summary in a timely manner. If an appropriate resolution to their inactivity is not feasible, they will be removed from the NAT. Members under the evaluation category working on other mapping-related projects may be moved to the structural category to better reflect their contributio` 로 업데이트 했기 때문에 번역도 따라서 수정했습니다.