Move Silence to Help_centre and fix lengths
Long silence lengths are now universally 43 hours, not 42 or (on older translations) 48 - except for the Russian version.

Some versions had the phrase "timed out" removed since it's not a phrase in osu!. Not all, however.

The Danish translation was partially fixed as well (it used to say users *themselves* could be removed after 3 months of a restriction instead of the restriction being removed).