Add note about reporting Ranked maps (#13585)
* Add note about reporting Ranked maps

It's always unclear how to report these.

The forum link is nearly invisible - I made it more visible. It might not be relevant anymore, though, so consider removing it.

I also fixed the issue in a previous translation where "References" wasn't translated.

* Add point to `Special Cases` instead

Also cleans up other parts of the page.

* Apply Hivie and MCheca's changes

Co-authored-by: Hivie <rayenattia@gmail.com>

---------

Co-authored-by: Hivie <rayenattia@gmail.com>
Co-authored-by: Hivie <me@hivie.tn>