Archive osu! Completionists
The badge has not been awarded to any users since despite multiple users meeting the criteria. Not even users like osu420 for which malicious intentions were not suspected by anyone have failed to get the badge on osu!std. Archiving the badge is best to avoid scenarios where users complete a ruleset and are told at the very end that they don't deserve it.