Update file storage paths in `Client/Release_stream/Lazer/File_storage` (#13882)
A user complaint that occurs much more than I would like to see it
happen is that someone backs up just the `files` directory inside the
lazer user folder, which *is not enough* to recover from, as
`client.realm` is also required.

There's even a loud `IMPORTANT READ ME` file in the lazer user folder
that even says the same thing, and yet it still keeps happening.

Today I randomly stumbled upon this wiki article again and noticed it
was pointing to the `files` directory, which *may* explain why this is
happening so often - people might be googling "how to move lazer" or
whatever, finding this article, which results in an incorrect conclusion
and an undesirable outcome.

And yes, before you ask: the information previously in the article was
*technically* correct, but it could also be harmful. And we *could*
address this in client via fifteen disparate ways people have proposed,
but addressing this in client via those ways is inherently more
time-consuming and risky than a wiki article update, because a wiki
article update doesn't risk *absolutely every lazer user losing their
data* if we mess up whatever migration process is used to hopefully
"address" this problem.