Clean up and reorganise `Handling foul play`
I have made `Handling foul play` focus on actions that violate competitive integrity. If a user wonders why they are restricted, they should consult `Account restrictions`.

I added IDs to the long section names. The addition of the continual misbehaviour clause to `Account restrictions` is because that makes much more sense there than on `Handling foul play`.'

The update to `Silences` is because the end of the article seems more fitting there. Line 65 is very much up to discussion, however.